pyproject.toml
[tool.poetry]
name = "task-duration"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "task_duration"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
